Move low-level shapefile code to impl. (#221)
This is a big merge. There are now geography-by-geography unit tests all across the on-
and off-spine geographies. See `test_geo_integration.py`. There is also a new utility for
scanning the U.S. Census shapefile offerings. It is in `scanshp.py`. It should be extended 
in and will be helpful for #172.