Fix DeprecationWarning in add_inferred_geography
It appears that the recommended way to fix this warning is to set `include_groups=False` (see https://stackoverflow.com/a/78100669). However, that only worked if I changed `group_keys` to be `True`.

This led to the resulting df having the correct data but in the wrong form (the rows were in the wrong order, and there was now a multi-index). The calls to sort_index and reset_index get the resulting data in the same form as the test data.

I am still fairly new to Pandas and find the way it handles indexes and group bys to be confusing. While this PR successfully removes the warning and passes the test, I welcome feedback that makes it more "pythonic".