Fix pandas warning that happens during tests
Specific warning was:

tests/test_geo_integration.py::AddInferredGeographyTestCase::test_tract
  /Users/arilamstein/censusdis/censusdis/data.py:1749: DeprecationWarning: DataFrameGroupBy.apply operated on the grouping columns. This behavior is deprecated, and in a future version of pandas the grouping columns will be excluded from the operation. Either pass `include_groups=False` to exclude the groupings or explicitly select the grouping columns after groupby to silence this warning.
    .apply(

The include_groups=False suggestion does not work here (unlike the other case I just fixed) because the column we're grouping by needs to be accessed by `add_geography`, which we're calling in the lambda. So I did the second option recommended in this SO: https://stackoverflow.com/a/78100669. I.e. set the index prior to the groupby.

Technically I could have just set removed the `group_keys` parameter (as opposed to setting it to `True`, its default value). But I thought it would make the diff more explicit. Also, `sort=False` is not strictly needed here because there are no tests that fail due to sorting (unlike the previous issue I fixed). But I thought it was better for the sake of consistency to do that here too. And it's probably better for the client to preserve whatever order they give us for the rows in their df.