Fix pandas warning during test
Specific warning is:

tests/test_plotspec.py::PlotSpecTestCase::test_plot
  /Users/arilamstein/censusdis/censusdis/maps.py:792: DeprecationWarning: DataFrameGroupBy.apply operated on the grouping columns. This behavior is deprecated, and in a future version of pandas the grouping columns will be excluded from the operation. Either pass `include_groups=False` to exclude the groupings or explicitly select the grouping columns after groupby to silence this warning.
    gdf = gdf.groupby(gdf[group_column], group_keys=False).apply(

Passes all tests. It's worth noting that while `pytest` still has some warnings, and some are geopandas related, this appears to be the last of the Pandas-specific ones, so we should at least consider closing #282.

This follows the most common fix I found to the warning: explicitly use the group_keys while doing the groupby, do not sort the df (so preserve the row order you were given), do not include the groups in the apply, and then reset the index to what we had at the start.